Final Project/Python/requirements.txt
blinker==1.8.2
click==8.1.7
contourpy==1.2.1
cycler==0.12.1
filelock==3.14.0
Flask==3.0.3
fonttools==4.51.0
fsspec==2024.5.0
importlib_metadata==7.1.0
importlib_resources==6.4.0
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.9.0
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.4
packaging==24.0
pandas==2.2.2
pillow==10.3.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
scikit-learn==1.4.2
scipy==1.13.0
seaborn==0.13.2
six==1.16.0
sympy==1.12
threadpoolctl==3.5.0
torch==2.2.2
torchvision==0.17.2
tqdm==4.66.4
typing_extensions==4.11.0
tzdata==2024.1
Werkzeug==3.0.3
zipp==3.18.2
